fix(effects-shared): IT smooth macro slide in float32
C stores channel_data.macro val/target/slide as float (player.h:251-253)
and accumulates val += slide each tick in float32. Our doubles drifted
by 1 ULP, landing (int)val one apart from C. Round the slide computation
(effects.c:834) and each accumulation (player.c:505) with Math.fround.

fixes it_smooth_macro

+++ packages/effects-shared/src/process.ts
@@ -596,7 +596,10 @@ function processRest(
       if (core.ctx.p.speed !== 0 && xc.macro.val < 0x80) {
         SET(xc, VolSlideFlag.MIDI_MACRO);
         xc.macro.target = fxp;
-        xc.macro.slide = (fxp - xc.macro.val) / core.ctx.p.speed;
+        // C: slide is float32 ((float)fxp - val) / speed (effects.c:834).
+        xc.macro.slide = Math.fround(
+          Math.fround(fxp - xc.macro.val) / core.ctx.p.speed,
+        );
       }
       break;
   }
